Authorization Policy Implementation

+++ app/Services/ProjectService.php
@@ -7,6 +7,7 @@ use App\Http\Resources\Project\ProjectResource;
 use App\Http\Resources\Task\TaskResource;
 use Illuminate\Contracts\Queue\EntityNotFoundException;
 use Illuminate\Http\Resources\Json\AnonymousResourceCollection;
+use Illuminate\Support\Facades\Gate;
 
 class ProjectService {
     public function __construct(private readonly ProjectRepository $repository) {}
@@ -19,7 +20,11 @@ class ProjectService {
         if(!$this->repository->existsByColumn('id', $id)) {
             throw new EntityNotFoundException('Project', $id);
         }
-        return new ProjectResource($this->repository->show($id));
+
+        $project = $this->repository->show($id);
+        Gate::authorize('show', $project);
+
+        return new ProjectResource($project);
     }
 
     public function update(int $id, mixed $data): ProjectResource {
@@ -28,6 +33,7 @@ class ProjectService {
         }
 
         $project = $this->repository->show($id);
+        Gate::authorize('update', $project);
 
         $user = auth()->user();
         $data['user_id'] = $user->id;
@@ -45,6 +51,10 @@ class ProjectService {
         if(!$this->repository->existsByColumn('id', $id)) {
             throw new EntityNotFoundException('Project', $id);
         }
+
+        $project = $this->repository->show($id);
+        Gate::authorize('destroy', $project);
+
         $this->repository->destroy($id);
     }
 
@@ -54,6 +64,8 @@ class ProjectService {
         }
 
         $project = $this->repository->show($id);
+        Gate::authorize('tasks', $project);
+
         return TaskResource::collection($project->tasks);
     }
 
@@ -63,6 +75,8 @@ class ProjectService {
         }
 
         $project = $this->repository->show($id);
+        Gate::authorize('tasksByStatus', $project);
+
         return $project->taskStatus()->with('tasks')->get();
     }
 }
