+++ app/Services/TaskService.php
@@ -7,6 +7,7 @@ use App\Http\Resources\Comment\CommentResource;
 use App\Http\Resources\Task\TaskResource;
 use Illuminate\Contracts\Queue\EntityNotFoundException;
 use Illuminate\Http\Resources\Json\AnonymousResourceCollection;
+use Illuminate\Support\Facades\Gate;
 use Illuminate\Support\Facades\Log;
 
 class TaskService {
@@ -17,14 +18,22 @@ class TaskService {
         if(!$this->repository->existsByColumn('id', $id)) {
             throw new EntityNotFoundException('Task', $id);
         }
-        return new TaskResource($this->repository->show($id));
+
+        $task = $this->repository->show($id);
+        Gate::authorize('show', $task);
+
+        return new TaskResource($task);
     }
 
     public function showComments(int $id): AnonymousResourceCollection {
         if(!$this->repository->existsByColumn('id', $id)) {
             throw new EntityNotFoundException('Task', $id);
         }
-        return CommentResource::collection($this->repository->show($id)->comments);
+
+        $task = $this->repository->show($id);
+        Gate::authorize('showComments', $task);
+
+        return CommentResource::collection($task->comments);
     }
 
     public function store(mixed $data): TaskResource {
@@ -37,6 +46,8 @@ class TaskService {
         }
 
         $mainTask = $this->repository->show($id);
+        Gate::authorize('storeSubtask', $mainTask);
+
         $subTask = $this->repository->store($data);
 
         $mainTask->subtasks()->attach($subTask->id);
@@ -50,6 +61,7 @@ class TaskService {
         }
 
         $task = $this->repository->show($id);
+        Gate::authorize('getAllSubtasks', $task);
 
         return TaskResource::collection($task->subtasks);
     }
@@ -59,6 +71,8 @@ class TaskService {
             throw new EntityNotFoundException('Task', $id);
         }
         $task = $this->repository->show($id);
+        Gate::authorize('update', $task);
+
         return new TaskResource($this->repository->update($task, $data));
     }
 
@@ -66,6 +80,9 @@ class TaskService {
         if(!$this->repository->existsByColumn('id', $id)) {
             throw new EntityNotFoundException('Task', $id);
         }
+        $task = $this->repository->show($id);
+        Gate::authorize('destroy', $task);
+
         $this->repository->destroy($id);
     }
 }
